fluxion: Prefix the module names with fluxion
Problem: Need to easily tell the difference between
the simple scheduler in flux-core with our graph-based
scheduler. Need to have a distinct name for future
publications of our technology.

Add "fluxion" as the prefix to both resource and
qmanager modules.

Change the RPC service names in fluxion-resource
to be prefixed with "fluxion-" as well.

Change the rc1 and rc3 scripts names for these
modules.

Adjust all of the test cases with the new module
names.

Update the main README.md accordingly.