resource: Use correct typedefs
Change the use of ggv_t to vtx_t. ggv_t is the data type of
a vertex of a graph generation recipe graph whereas vxt_t
is the data type of a vertex of the actual resource graph.

Found a few sites where ggv_t type is used when vtx_t
should have been used during a code review and fix them.