writer: Drop pretty print support for JSON based writers
Problem: It is becoming difficult to maintain two difference
versions (pretty vs. unpretty) of writers of the same type.
Further the value of the pretty versions is marginal as the
output from a unpretty writer can easily be made
human readable by piping to the jq command.

Drop the pretty writer support for JSON-based writers.

Adjust testing coverage accordingly.

Will consider to add back this support when we convert
our JSON writers to use either the Jansson library
or its C++ equivalent.

If and when this is revived, the implementation should be added
such that pretty writers will be supported without
having to introduce their own classses that contain
duplicate codes with unpretty versions.