writer: JGF writer now emits paths for each vertex
Problem: The paths field was previously omited
with a hope that the paths can be reconstructed
by the future JGF reader.
But it appears that this would require lots of code
at the reader side and would be inefficient regardless.

Modify the JGF writer code to emit the paths of each
graph vertex within each subsystem.
(Doing this in the writer side is simpler
at the expense of using more memory/storage).

Will consider JGF schema compression as part of
future optimization phase.