writer: avoid condense encoding of edge relationship
Problem: The old scheme is to concatenate the edge's
relationship information with respect to each subsystem
using some delimiters. While this is condense, this
can lead to unnecessary complexity in the future
JGF reader code.

Emit the relationship information for each subsystem
that the edge represents as a unique sub-key within
the name key.