reader: introduce the resource reader factory class
Instantiate an object of a derived resource reader class
that corresponds to the input string (reader type name)
and return it as the object of the resource reader base
class.

Allow the client code to be written as the base class,
making it future proof.

Use shared_ptr class to enhance safety.