resource: add performance variation-aware match policy
The policy relies on the use of the "perf_class" property for the resource.
This "perf_class" is provided by the user through the set-property interface.
The range for perf_class is 0 through 9999.
A lower value indicates better efficiency of the resource.
As a result, fold::less is used for scoring.
In the case where there is a missing "perf_class", we assume the worst
efficiency, which we designate as "perf_class" of 9999.

style: fix whitespace for stoi in dfu_match_var_aware.cpp.