resource: test commands for variation-aware policy.
There are 5 tests for resource-query. These are the test command files.
Tests 1-4 have 4 jobs and test the match allocate.
Test 1 checks for correct inputs and verifies the allocations.
Tests 2,3,4 test for missing, invalid, and out of range inputs for perf_class.
These three tests assume the worst-efficiency for the resources that have
incorrect input, and thus use a perf_class of 9999 for them.

Test 5 has 8 jobs and tests for match allocate-orelse-reserve.