c++: first-cut C++ interface for flux_msg_handler
This adds flux.hpp in libutil (for now) that wraps around the callback
and arg specified by the user to create an exception-safe callback which
is then passed on to flux.  It also has defaulted arguments for the most
common values of typemask, rolemask and matchtag, so one interface can
be used instead of two.  With uniform initialization, this makes the C++
interface look a great deal like addvec, as in resource_match.cpp
setting up all of the handlers and scoping them with RAII is just this:

```c++
flux::msg_handler_wrapper handlers[] = {
    {h, "resource.match", match_request_cb, h},
    {h, "resource.cancel", cancel_request_cb, h},
    {h, "resource.info", info_request_cb, h},
    {h, "resource.stat", stat_request_cb, h},
    {h, "resource.next_jobid", next_jobid_request_cb, h},
    {h, "resource.set_property", set_property_request_cb, h},
    {h, "resource.get_property", get_property_request_cb, h}
};
```

There is more work to be done on propagating more meaningful information
out of exceptions, and being able to take an arbitrary functor from C++,
but this is the basic building block we can start from.