resource: add virtual in-edge support for root vertices
Problem: The traverser uses graph edges to annotate
selection information such as how much resource
in their target vertices can be selected. This info
is then later used by update() to finalize the best
matching subgraph. Now, because the root vertices
of the graph do not have in-edges, this scheme required
special casing for the roots.

Add virtual in-edge support for root vertices so
that the core selection/update algorithm does
not have to have special casing.

Fix one of the side effect of the special casing:
longer method parameters for core methods like
select () and update ().