reader: add update() support
Add update() to the base class as well as
its derived classes.

Implement update() within the JGF reader class.
Other derived classes get an empty update()
implementation for future extension.

- Refactor the JGF parser code for the unpack()
method so that it can be reused for the update()
operation as well.

- Find the vertices and edges corresponding to
the subgraph written in JGF.

- If the vertex is marked "exclusive", update
its schedule.plans planner object as well as
schedule.reservations and schedule.allocations
table. Note that we update these data in this
reader code, not in the traverser code, so
even those vertices that the traverser won't
walk still get their plans updated correctly.

- update_edges() updates the source edge
and target edge of each JGF edge element. It
adds selection information such as how much
resource in the target vertex must be selected
to the edge object which then later gets used
by the traverser code's update() method.