traverser: integrate the update method of readers
Add support so that upper layer can update the existing
resource graph using a JGF subgraph.

Overload run() method so that the traverser can run
to update the resource graph using a JGF subgraph,
jobid, starttime and duration as well as a reader
object. Note that only the JGF reader supports this.

Significantly refactor and clean up the traverser's
update walking code so that it can be used both by
the existing run() as well as this alternative
run() method.

Pass the resource_graph_db_t object as well as
the filter graph object to handle the cases
where it must update or cancel a JGF subgraph
created by a different traverser instantiation.
For example, this condition can arise when the
resource module is is reloaded with a different
match policy and the job-manager wants
to reconstruct a previous allocated job.