resource: fix buffer overflow from issue #509 Copying from Dong's comment: We treat slot in a different way than any other resource type because it is the only non-physical resource type. How we handle is this is : 1) perform a DFV subtree walk for a slot type, 2) divide the resource set discovered from this DFV walk equally according to the slot shape, 3) check if the given number of slots can be satisfied. There turns out to be a bug in 2).
For example, if your jobspec is
socket[1]->slot[1]->core[2] and your
machine is socket[1]->core[8], then at the
end of the DFV walk on the subtree rooted
at a socket vertex, we will give you core: 8.
Then, we divide 8 by 2 which means that we
can have 4 slots. So clearly this satisfies
slot[2].

Now, there is an issue in the current code.
The code doesn't factor into account the
granularity in which a resource pool vertex
is modeled/constructed for this.

Say, each memory pool granularity is 64GB and
your machine has 4 of those. In terms of quantity,
your machines has socket[1]->memory[256].
But in terms of scheduleable units you only
have 4 (i.e., 4 x 64GB). Now if you jobspec
is socket[1]->memory[32], we will currently
say you have 8 slots when in fact you only
can fit 4 slots because of this granularity
constraint! This leads to a buffer overflow.

These changes fix the resulting segfault.