test: add t3017-resource-granule.t
Add two subtests within t3017-resource-granule.t that checks
the correctness of matching when each memory resource vertice
is modeled as a pool (of size 2GBs).

Even though each jobspec requests 1GB of memory, because
the finest granularity of memory to be scheduled is 2GBs,
you can schedule only 4 jobspecs.