t/scripts/flux-resource: update exit codes
Problem: exit codoes from the test flux-resource script do not appear
to be consistent between Python 2.7 and 3.6. For example, EBUSY errors
from the tool are being thrown as IOError instead of EnvironmenError
under 3.6.

Consolidate the handling of IOError and EnvironmentError so that exit
codes are based on errno instead of the exception class. This should
make exit codes consistent between Python versions.