testsuite: update usage of flux-job list in tests
Problem: In upstream flux-core the `flux job list` plumbing tool
was changed to only output JSON, which breaks some of the flux-sched
tests which use this tool to count active and running jobs.

Update flux-job list usage in the qmanager tests to prevent failures.

Note: although `flux job list` and `flux job list --states=active`
are currently equivalent, use the latter form in tests to be
explicit as well as protect the tests from potential changes
in the default output of `flux job list.

Fixes #570