flux-tree: add command to easily create a Flux instance hierarchy
Create a Flux instance hierarchy according to the specified
policies and schedule/run the specified number
of Jobscripts at the last level of this hierarchy.

If --topology=2x4 and --njobs=32 are given, for instance,
2 Flux instances will be spawned from within the current instance,
each of which will in turn spawn 4 child Flux instances, totaling
8 instances at the last level of this hierarchy.
Once this is done, 4 jobs (of Jobscripts) will be scheduled
and executed at each of these 8 last-level Flux instances.

The resources specified by --nnodes and --ncores-per-node
are recursively divided such that each sibling Flux instance
will be assigned to an equal split of the resources of their
parent instance.

Jobscript is expected to submit one or more programs through
the flux-job submit command or its variants.
Jobscript is passed with four environment variables:
FLUX_TREE_ID, FLUX_TREE_JOBSCRIPT_INDEX, FLUX_TREE_LEAF_NNODES
and FLUX_TREE_LEAF_NCORES_PER_NODE.
FLUX_TREE_ID is an ID string uniquely identifing the hierarchical
path of the Flux instance on which Jobscript is being executed.
FLUX_TREE_JOBSCRIPT_INDEX is the integer ID of each jobscript
invocation local to the Flux instance. It starts from 1 and
sequentially increases.
FLUX_TREE_LEAF_NNODES is the number nodes assigned to the instance.
FLUX_TREE_LEAF_NCORES_PER_NODE is the number of cores per node
assigned to the instance.

If --queue-policy (additionally --queue-params) and/or
--match-policy are given, each level of this hierarchy will
be set to the specified queuing and matching policies and
parameters. Otherwise, all levels will be configured
to be used either the default policies or policies specified
through the FLUX_RESOURCE_OPTIONS and/or FLUX_QMANAGER_OPTIONS
environment variables.

If any one of Jobscripts returns a non-zero exit code, flux-tree
detects the script invocation exited with the highest code and print
both that exit code and the outputs printed from executing the script.
In this case, FLUX_TREE_ID and FLUX_TREE_JOBSCRIPT_INDEX are also
reported in the from of
${FLUX_TREE_ID}@index[${FLUX_TREE_JOBSCRIPT_INDEX}]