testsuite: avoid rc3 hang in qmanager-reload test
Problem: flux-core rc3 now tries to cancel running jobs and
wait for them to go through cleanup.  This results in a hang
when rc3 is not run under a timeout.

This works at the moment because rc3 quickly times out when
the shutdown grace period expires, and such timeouts are not
treated as instance errors.  flux-framework/flux-core#2733
proposes having the grace timeout NOT cover rc3, and handling
rc3 errors as instance errors.

Reload sched-simple on test exit so that rc3 can successfully
clean up running work.