README: clarify when to use `--prefix` when installing flux-sched
the instructions for install flux-core in the flux-core README do not
use the `--prefix` flag, but the install instruction for installing
flux-core and flux-sched in the flux-sched README direct users to
use `--prefix=$HOME/local`.  If the user followed the install
instructions for flux-core in flux-core's README and the install
instructions for flux-sched in flux-sched's README, they would end up
with the two projects installed in different locations

reference the flux-core install instructions, rather than duplicating
them

update the flux-sched README to more explicitly describe what to do in
the flux-core system-installed vs side-installed cases.

add a note about the convenience of having flux-core and flux-sched
installed to the same prefix.