conf: add a configuration file for qmanager
Add qmanager.toml as the config file for
the qmanager module.

Introduce a config hierarchy for qmanager
in this file, the hierarchical keys are:
qmanager, qmanager.queue-params, and
qmanager.policy-params.

One and only one key is supported
at the top level (qmanager) to specify
the queueing policy to use: "policy".

At the qmanager.queue-params level, two keys
are supported to control the depth of
pending jobs in the queue: max-queue-depth
and queue-depth. Currently, they will be
applied to all of the queuing policies.

At the qmanager.policy-params level, two keys
are supported for conservative and/or hybrid
backfill policies: max-reservation-depth
and reservation-depth.
Note that reservation-depth is only applied to
"hybrid" as the reservation depth of the
conservative policy must be the max value
by definition.