build: set FLUX_PYTHON_VERSION in AX_FLUX_CORE
Set FLUX_PYTHON_VERSION in AX_FLUX_CORE macro and set the
default for PYTHON_VERSION to this value if not set by the user.

This allows flux-sched configure to pick the same version of Python
as flux-core was configured with without requiring the user to
set PYTHON_VERSION manually.