rc: use reload -f to load qmanager/resource
Problem: Some of the in-tree tests are failing when
flux-sched was already installed with flux-core.
Their rc1 scripts are executed and load those
"installed" modules instead of "in-tree" modules.

Use "flux reload -f <qmanager|resource>" so that
when these failing tests load them via
FLUX_RC_EXTRA, the in-tree versions are loaded.