flux-tree: add helper for scalably collecting performance data
Problem: constructing and manipulating JSON blobs via the command line
is slow and quickly exceeds the limit for command line argument length
set in the linux kernel.

Solution: Add a helper written in python that performs the exact same
action but using python's standard library utilities and the flux python
bindings.  This reduces the number of processes launched, eliminates the
need for using command line arguments to pass JSON blobs between
programs, and reduces the total number of times the JSON needs to
be (de-)serialized.

Results:
Before the helper, for a tree topology of 32x36, the perf
results record a runtime of 45 seconds, but the `time` utility used on
the flux-tree command directly reports a runtime of 3 minutes 45
seconds.  After the introduction of the helper, for the same tree
topology, the perf results record a runtime of 34 seconds, and the
`time` utility reports a runtime of 47 seconds.  So the helper not only
speeds up the post-processing time, but it also has a measurable impact
on the reported results as well.

Before the helper, for a tree topology of 1x32x36, the root instance in
the tree topology (i.e., not the instance that flux-tree is directly run
inside of), would hit the linux kernel limit for command line argument
length whilst processing the JSON performance data. This would cause the
script to error out when collecting performance results and fails to
write any performance data out to the filesystem. After the helper, the
same tree topology runs to completion, successfully saves out the
performance data, and does so with a total runtime (as reported by
`time`) of 53 seconds.