flux-tree: add --flux-rundir support
Problem: For a large flux-tree test where lots of flux
instances are used, the content store back-ends can
use lots of memory leading to OOM.

Provide --flux-rundir=DIR option which sets the
rundir to DIR so that the content stores can be
redirected to a subdirectory in DIR for this case.

Use the rundir attribute because upon the exit
of each Flux instance, its content store sqlite
file is removed.