resource: Add the resource.update RPC
Problem: to allow qmanager to reconstruct
the resource states of running jobs on
reload in the future, the resource module
needs a way to update its resource graph
using the R(V1) of each running job.

Introduce a new RPC (resource.update) that
can update the resource graph using the
R(V1) string sent to it.

Call the dfu_traverser_t class' overloaded
run() method to perform this update.
To satisfy it's inputs, parse out the JGF
section as well as the starttime and
expiration keys from the R.

This is different from the existing
resource.match RPC that finds the best
matching resources according to the jobspec
and update the resource graph with those
matching resources.

Return the previously recorded R data
if the requested jobid already exists
and the requested R is equal to the
recorded R of that jobid. This will be
the case if only qmanager was reloaded,
not resource.

Return a corresponding errno
if the requested jobid already exists
and the requested R is not equal to
the recorded R or any other error
condition arises.