queue policy: rename reconstruct into reconstruct_queue
Problem: The semantics of reconstruct within the queue policy
implementation class is to reconstruct only the running
job queue.  The method name, reconstruct(), hence is
misleading as it may suggest the method would reconstruct
both the queue *and* resource state.

Change the name to be "reconstruct_queue" to clarify this.
Add more error checking code as part of this as well.