queue-policy: APIs to reconstruct the resource state
Problem: we have an API to reconstruct the state
of running-job queue state for a job, but we don’t have
an interface to reconstruct the state of resource
data store for the job.

Add reconstruct_resource as a pure virtual method
to the base queue policy class (queue_policy_base_t).
Because the base queue policy classes (queue_policy_base_t
and queue_policy_base_impl_t) must manage various job
queues while being agnostic to the resources of jobs, this
pure virtual method must be implemented by the
derived classes, the queue policy layers that aware
of the resources. Still, this virtual interface allows
queue_policy_base_t to make a up-call to the methods
overridden by all of the derived classes. This makes
this scheme future proof with respect to future
queueing-policy implementation classes.