qmanager: Add basic state recovery support
Implement the basic state recovery semantics
using the recent changes made to the resource
module, resource high level API, queue
policy layer.

Recover the state of both the running job
queue and the resource graph data store
for those running jobs.

Meet the state-recovery requirement for a
key near-term use case in which both
resource and qmanager modules are shut down
and reloaded in a coordinated fashion.

Assume two conditions for full recovery:

1. The RV1 JSON object for each and every
   currently running job includes the
   "scheduling" key, therefore the JGF
   section.
2. The resource module is reloaded before
   qmanager.

Lay some foundational work to prepare for
more comprehensive failure recovery
in the future.

If only the qmanager module is shut down
and reloaded, we still fully recover
the state.

If both qmanager and resource
are reloaded, but the RV1 JSON
of a running job does not include
the "scheduling" key, qmanager
will fail to load because
it can't recover the resource state.
However, if "resource-recovery-on-load=false"
is given as a qmanager load option,
we make the following trade off:

+ qmanager will load and schedule
+ qmanager will allow clean FREE
  of the running jobs when they complete
  (i.e., those jobs that were already
   running on reload).
- but the resources that those running jobs
  are using can be scheduled to other
  jobs (i.e., overscheduling) because
  the resource module won't be able
  to reconstruct the state.

To summarize the current state recovery semantics:

      module       |   RV1 of running jobs
resource  qmanager | JGF   no-JGF
--------------------------------------------------
reload    reload   | full  partial*
running   reload   | full  full
reload    running  | no    no

* Footnote: resource-recovery-on-load=false option
can still allow qmanager to be loaded but the
resource state will not be recovered.