build: don't install qmanager.toml
Problem: installing qmanager.toml might overwrite a user's
config file in an unpackaged environment.

Don't install the file.

In addition, flux-core was changed in flux-framework/flux-core#2866
to _only_ use config files if a path is specified on the command line
or environment.  The default config file directory is no longer defined.