testsuite: force in-tree qmanager config testing
Problem: an installed qmanager could fail on test input
before workaround to unload and reload the in-tree qmanager
can be executed.

Change the test to use purpose-built rc1/rc3 scripts to
load qmanager and minimal dependencies, ensuring that the
in-tree module is loaded, and setting command line arguments
as specified in QMANAGER_OPTIONS environment variable.

Pull common code out to shared shell functions, so that it doesn't
need to be modified in every test.