flux-tree: modify its helper to run w/o a Flux instance
Problem: When flux-tree is run in dry run mode, it is not
assumed to be run under an active Flux instance.
But flux-tree-helper returns the ENOENT error
under this condition.

Assign None to flux_handle when flux.Flux() fails to
connect to a Flux instance. Other calls that require
a valid flux_handle is guarded with the
flux_handle != None check.

Assume the num_children argument must also be zero
when flux_handle != None.