testsuite: fix potential hang in t1002-qmanager-reload.t
Problem: The t1002-qmanager-reload.t test will hang during rc3 when
used against a Flux where sched-simple acquires resourcess from the
resource module. This occurs because the test reloads hwloc XML
out of band, and the core resource module (from which sched-simple
gets its resource list) has a different resource configuration.
This causes sched-simple to reject all free requests as invalid
at the end of the test, causing the wait for queue drain to block
forever.

In order to fix this, the resource module must be reloaded, which
in turn requires that sched-simple be reloaded, after hwloc XML is
overwritten.