testsuite: add test coverage for full recovery cases
Add cases to test flux-sched modules can fully
recover the state of running jobs on reload
for the cases where they must:

1. The format of running jobs at recovery point
   is rv1 and we restart both resource and
   qmanager modules.

2. We restart only qmanager, not
   the resource module.

Note that we also added test coverage for a condition
where a running job completes
after fluxion modules are unloaded, but before they are
reloaded. In test 5, job1 is cancelled after
both of the fluxion modules are unloaded.
Then, the following test is augmented to verify that job1
is not enqueued on reload.