t/mt-storage: add storage-constrained tests for node-local storage
Problem: previous multi-tier storage tests allocated compute and storage
in the exact same ratio as found in the cluster.  It is conceivable that
these tests would pass under a scheduler that only respected compute
constraints while ignoring IO constraints.

Solution: add tests with jobspecs that request 2x storage of the
existing tests to exhaust the storage resources before the compute ones

Note: the node-local storage case is known in the tests as L0