configure: use newer tar format to support longer filenames
Problem: when running `make distcheck`, the creation of a tarball fails
due to filenames that are too long.  This is because autotools defaults
to the oldest, most compatible tar format (v7) which limits filename
length to 99 characters.

Solution: switch to a newer tar format (which is honestly still quite
old - defined in the 1988 POSIX Standard) called `ustar` that supports
filenames up to 256 characters in length. If this becomes a problem in
the future, we can switch to `pax` which has no limits.  Additionally,
set the filename-length-max option so that the entire build fails rather
than the tar silently failing to be created.

Further Reading:
- https://www.gnu.org/software/tar/manual/html_section/tar_67.html
- https://www.gnu.org/software/automake/manual/html_node/List-of-Automake-options.html
- https://noiselabs.io/tar-file-name-is-too-long-max-99/