t/scripts/*.py: add a bang line
Problem: the python scripts in `t/scripts` are intended to be run as
subcommands via the `flux` binary, and the comments at the top of the
file make that clear.  Unfortunately, if you decide to execute them
directly anyway (e.g., `./flux-ion-resource.py`), they get interpreted
as bash scripts.  In one unfortunate case, they locked up my entire
machine (still not sure why that happened).

Solution: add a bang line that points to `/bin/false` so that when they
are directly executed, they immediately fail with an exit code of 1.