resource: Add resource matching expression eval class
Problem: We will soon need find() within the top-level
depth-first-and-up (DFU) traverser class, which
will take in a matching criteria and walk
the resource graph in DFU order to find all
of the matching resource vertices
and emit the found resources.
As a first step, we need an evaluator
that will parse an expression and evaluate it
against the state of each vertex during the walk.

Introduce the expr_eval_t class that can
validate or evaluate a matching criteria:

  - Each individual matching criterion (a.k.a. leaf
    expression) is expressed as a key-value pair,
    representing a predicate (p(x) where p is the key
    and x is the value).
  - Two predicates can be combined with "and"
    or "or" logical operators.
  - Parentheses are supported to group expressions
    with a higher operator precedence. For example, in
    expression
    "staus=up and (sched-now=allocated and sched-future=reserved)".
    The parenthesized expression is evaluated
    before taking the "and" operator with the
    the result of the first expression.

Introduce expr_eval_target_t, a pure abstract class
the target of the evaluator must inherit and implement.
Inheritance will easily adapt the target class
(e.g., resource_pool_t) to expr_eval_t in the future.