traverser: add find() into dfu_impl_t
Traverse the resource graph, evaluate
each visiting vertex and emit those resources
whose states are matched with the criteria
passed in.

Backgroud:

1. To evaluate each resource vertex, we view
each resource vertex as a set of predicates:
{ predicate1, predicate2, ..., predicateN }.

2. Example predicates are "is status down?" (status=down),
"is scheduled state allocated?" (sched-now=allocated).

3. Use the expr_eval_vtx_target_t adapter class
on a vertex so that we can evaluate
each queried predicate on its current state
without having to modify the resource vertex
classes themselves.

4. Because we allow the partial marking of a vertex,
the state of an ancestor vertex can often determine
the effective state of a child vertex.
For instance, if the status of a compute node vertex
is "down", the status of a core
underneath it must also be "down" even if its
actually marked status is still "up".
To support this case, we use the vtx_predicates_override_t
adapter class to aid in passing
ancestor predicates down to the offspring resources
throughout the traversal.

Add find() to our DFU traverser that takes in a writer
and match criteria. It performs the DFU visit on the
resource graph and conducts the following on
certain visit events:
  - On each preorder visit, it builds the predicates
of the visiting vertex and passes it to
the next vertex visit (depth first visit on
the dominant subsystem and up visit on
auxiliary subsystems).
  - On each postorder visit, it builds
expr_eval_vtx_target_t based on the state
of the visiting vertex *and* the predicates set
passed down from the parent and evaluates the
expression on this mixed state.

Note that the up-visit function definition is
introduced, but this function is currently
is NOOP returning 0.
This function will be implemented as part of a future
commit with an actual use/test case to help drive
our implmentation.

Also note that if at least one matching vertex
is found in a subtree rooted at a visiting vertex,
the visiting vertex is emitted even if it is not
evaluated to be true.
We include this vertex so that the emitted resources
will form a complete graph, which is currently required
by our match writers.

For RV2, we may want to introduce a way to differentiate
the actually matched vertices vs. those ancestor vertices
that are included to comply with the format.