query: add --test-output support for set|get-status
Problem: When a resource-query test is written
by redirecting stdout, a race can occur between
resource-query's prompt outputs (resource-query>)
and the testing output. This can lead to a false positive
when we compare the testing outputs with the expected
outputs.

Use the output file provided by --test-out (or -t)
(instead of stdout redirection) in order
to avoid a race.

Update t3024-resource-status.t to use -t to
use this feature.

Update this test's expected outputs (removing prompts
and the subcommands typed in).