t: run sierra2 hwloc XML through hwloc-convert and fix the fallout
** Problem **

The resource graph produced by the following process (referred to in
this commit message as "single read"):

- read V1-xml on disk with V2 API
- traverse hwloc topo with traversal APIs

is *NOT* equivalent to the graph produced by the following
process (referred to in this commit message as "double read"):

- read V1-xml on disk with V2 API
- serialize with V1 compatible writer
- read serialized V1-compatible XML with V2 API
- traverse hwloc topo with traversal APIs

The "single read" process, when applied to our `04-brokers-sierra2`
hwloc XML data, produces a resource graph where the GPUs are direct
children of the compute node. The "double read", process, when applied
to our `04-brokers-sierra2` hwloc XML data, produces a resource graph
where the GPUs are direct children of the sockets.  In terms of
locality, the latter is "more correct".  The difference in these two
graphs breaks matches against jobspecs that assume the GPUs are direct
children of the node; specifically `basics/test013.yaml`.

The "single read" process is what happens when you test with the
`resource-query` utility.  The "double read" process is what happens
when you test with the `flux ion-resource` utility against a
`fluxion-resource` module that has been populated with xml from `flux
hwloc reload`.  The "double read" process also more closely mimics what
happens "in production".

Note: All of the above "reads" use the following flags for the various
resource filtering functions added to V2:

  - io_types: HWLOC_TYPE_FILTER_KEEP_IMPORTANT
  - cache_types: HWLOC_TYPE_FILTER_KEEP_STRUCTURE
  - icache_types: HWLOC_TYPE_FILTER_KEEP_STRUCTURE

** Solution **

Run the `04-brokers-sierra2` XML files through the `hwloc-convert` test
utility in flux-core to emulate the first read in the "double read"
process.  Specifically it performs the first two steps of the process:
read V1-xml on disk with V2 API and serialize with V1 compatible writer.
The result is that `resource-query` and `flux ion-resource` now both
instantiate the same resource graph and thus produce the same results on
the same jobspecs and hwloc XML.

The resource graphs produced by these utilities now includes a socket in
between the nodes and GPUs, which affects a jobspec
request (basics/test013.yaml) and the expected output of two GPU-based
match tests (018.R.out and 021.R.out).  This commit updates the jobspec
and the expected outputs to include the socket resources.

Note: This commit does not attempt to normalize the resource graphs
produced by hwloc V1 versus V2.  As we discussed in
flux-framework/flux-sched#656, where this would cause issue in
production, we can leverage the use of `--load-allowlist` to filter out
resources that cause differences.  If one sticks strictly to Jobspec V1
and thus filters out `socket`s, this difference will be normalized out.