resource: satisfiability for resources going up/down
Problem: When a resource goes down and a job requests
all of the resources, the job spec is currently
evaluated to be unsatisfiable. This is incorrect.
In MATCH_ALLOCATE_W_SATISFIABILITY mode, we check
a job's satisfiability by trying to schedule
the job as late as possible. However, when a
resource is down, it doesn't matter how late we
try to schedule this: this job just cannot be
scheduled and our system will deem it unsatisfiable.
Similarly in MATCH_ALLOCATE_ORELSE_RESERVE mode, we
deem a job unsatisfiable when we iterate through
all of the scheduled points but find no scheduled
point that satisfies the request. Again, this
will not work when there is a resource that
is marked "down".

Solution:
Expand the allocation type within jobmeta_t beyond
boolean (allocation vs. reservation). If the allocation
type is satisfiability (AT_SATISFIABILITY), we just
ignore the current status of visiting resources
for matching so that those down resources will still
be considered for satisfiability checks.