writer: hoist emit_json() as a hard requirement
Problem: In the upcoming sched.resource-status
RPC, the emitted RV1 needs to be embedded into
a higher level JSON (e.g., under "all", "down"
and "alloc" keys). Converting the stringified
RV1 back to an JSON object just to embed it
into another JSON would be a waste in particular
when we already mostly have an JSON version
of the emit() method.

Solution:
Add emit_json() as another pure virtual
public method within the base writer class
so that all the derived writer classes are
required to implement it..

Add and implement this function for each
of the dervied writer methods.