traverser: ignore unknown rank for mark ()
Problem: A upper layer like sched-fluxion-resource
can pass those ranks that are not a part
of the resource set that has been grown.
But if mark() discontinues on a unknown
rank, this leads to a failure
to load sched-fluxion-resource.

Solution: Ignore those unknown ranks and
continue marking those resources
from known ranks.