testsuite: verify fluxion can handle resource exclusion
Verify Fluxion modules load well even if all the
rank resources are excluded. In this case, every
job will be rejected. Note this won't be that useful
until flux-resource will support "unexclude" as
a special case for resource growth.
We don't have this yet; so we don't test.

Verify Fluxion can correctly handle when two out
of four ranks are excluded.