qmanager: order pending jobs by priority and submit time
Problem: Currently the pending jobs within
fluxion's qmanager is ordered by enqueue time as seen
by qmanager, not by the enqueue time as seen
by job-manager. Further, it doesn't take
into account the job's priority either.
One problem this can cause is when a pending job
is modified. When modified, the existing sched.alloc
request will be cancelled and a new sched.alloc
request will be issued with an altered job spec.
However, with the current enqueue-time ordering,
this will put the newly modified job at the end
of the pending queue, which is incorrect.
Similary, with unlimited mode where all the submitted
jobs are enqueued into qmanager with no blocking
at the job-manager level, a high priority job
that is submitted later cannot be reordered to run
earlier.

Change the "key" of the pending queue map from
uint64_t to be std::vector<double> such that pending
jobs can be lexicographically ordered
according to this vector.

Use <priority, t_submit> as the current key.

This change can easily accomodate other
order-affecting changes that are expected
to be introduced in the future including
fair share priorities.